Se corrigen errores y se implementa manejo de exepciones

+++ src/controller.js
@@ -32,6 +32,20 @@ class LibroController {
   // Crear un nuevo libro
   async add(req, res) {
     const { nombre, autor, categoria, año_publicacion, ISBN } = req.body;
+    const validKeys = ['nombre', 'autor', 'categoria', 'año_publicacion', 'ISBN'];
+    const receivedKeys = Object.keys(req.body);
+
+     // Validar que solo se envíen los campos permitidos
+     const invalidKeys = receivedKeys.filter(key => !validKeys.includes(key));
+     if (invalidKeys.length > 0) {
+       return res.status(400).json({ error: `Atributos no válidos: ${invalidKeys.join(', ')}` });
+     }
+ 
+     // Validar que todos los campos obligatorios estén presentes
+     if (!nombre || !autor || !categoria || !año_publicacion || !ISBN) {
+       return res.status(400).json({ error: 'Faltan campos obligatorios' });
+     }
+
     try {
       const [result] = await pool.query(
         "INSERT INTO Libros (nombre, autor, categoria, año_publicacion, ISBN) VALUES (?, ?, ?, ?, ?)",
@@ -48,17 +62,20 @@ class LibroController {
   async update(req, res) {
     const { id } = req.params;
     const { nombre, autor, categoria, año_publicacion, ISBN } = req.body;
+  
     try {
+
+      const [existingBook] = await pool.query('SELECT * FROM Libros WHERE id = ?', [id]);
+      if (existingBook.length === 0) {
+        return res.status(404).json({ error: 'Libro no encontrado' });
+      }
+
       const [result] = await pool.query(
         `UPDATE Libros SET nombre = ?, autor = ?, categoria = ?, año_publicacion = ?, ISBN = ? WHERE id = ?`,
         [nombre, autor, categoria, año_publicacion, ISBN, id]
       );
-      if (result.affectedRows > 0) {
         res.json({ message: "Libro actualizado correctamente" });
-      } else {
-        res.status(404).send("Libro no encontrado");
-      }
-    } catch (err) {
+      } catch (err) {
       console.error(err);
       res.status(500).send("Error en el servidor");
     }
@@ -68,12 +85,15 @@ class LibroController {
   async deleteByISBN(req, res) {
     const { isbn } = req.params;
     try {
-      const [result] = await pool.query('DELETE FROM Libros WHERE ISBN = ?', [isbn]);
-      if (result.affectedRows > 0) {
-        res.json({ message: 'Libro eliminado correctamente' });
-      } else {
-        res.status(404).send('Libro no encontrado');
+      
+      // Verificar si el libro existe antes de eliminar
+      const [existingBook] = await pool.query('SELECT * FROM Libros WHERE ISBN = ?', [isbn]);
+      if (existingBook.length === 0) {
+        return res.status(404).json({ error: 'Libro no encontrado' });
       }
+      
+      const [result] = await pool.query('DELETE FROM Libros WHERE ISBN = ?', [isbn]);
+      res.json({ message: 'Libro eliminado correctamente' });
     } catch (err) {
       console.error(err);
       res.status(500).send('Error en el servidor');
